refactor(suite): consolidate firmware components and enhance installation flow
