--- packages/suite/src/components/firmware/CheckSeedStep/FirmwareInstallationBackupButton.tsx
@@ -1,31 +1,0 @@
-import { Button } from '@trezor/components';
-
-import { goto } from 'src/actions/suite/routerActions';
-import { Translation } from 'src/components/suite/Translation';
-import { useDispatch } from 'src/hooks/suite';
-import { useFirmwareDesktopUpdate } from 'src/hooks/suite/useFirmwareDesktopUpdate';
-
-type FirmwareInstallationBackupButtonProps = {
-    isBackedUp: boolean;
-};
-
-export const FirmwareInstallationBackupButton = ({
-    isBackedUp,
-}: FirmwareInstallationBackupButtonProps) => {
-    const dispatch = useDispatch();
-
-    const { resetReducer } = useFirmwareDesktopUpdate();
-
-    const secondaryButtonText = isBackedUp ? 'TR_CHECK_SEED' : 'TR_CREATE_BACKUP';
-
-    const handleSecondaryButtonClick = () => {
-        resetReducer();
-        dispatch(goto(isBackedUp ? 'recovery-index' : 'backup-index'));
-    };
-
-    return (
-        <Button variant="tertiary" onClick={handleSecondaryButtonClick}>
-            <Translation id={secondaryButtonText} />
-        </Button>
-    );
-};
