+++ packages/suite/src/components/firmware/FirmwareInitial.tsx
@@ -1,79 +1,16 @@
-import { useState } from 'react';
-
-import styled from 'styled-components';
-
-import { ExtendedMessageDescriptor } from '@suite-common/intl-types';
-import { AcquiredDevice } from '@suite-common/suite-types';
-import { selectDevices } from '@suite-common/wallet-core';
-import { ButtonProps, Column, Note, Row, Tooltip, variables } from '@trezor/components';
+import { getFwUpdateVersion } from '@suite-common/suite-utils';
+import { Banner, Card, Column } from '@trezor/components';
 import { FirmwareType } from '@trezor/connect';
-import { DeviceModelInternal, isBitcoinOnlyDevice } from '@trezor/device-utils';
-import { spacingsPx } from '@trezor/theme';
+import { getFirmwareVersion } from '@trezor/device-utils';
 
-import { FirmwareOffer, FirmwareWarningsList } from 'src/components/firmware';
-import { OnboardingCard } from 'src/components/onboarding/OnboardingCard/OnboardingCard';
-import { SkipStepConfirmation } from 'src/components/onboarding/SkipStepConfirmation';
+import { FirmwareOffer, FirmwareWarningsList, FirmwareWipeWarning } from 'src/components/firmware';
 import { Translation } from 'src/components/suite/Translation';
-import { useDevice, useOnboarding, useSelector, useTranslation } from 'src/hooks/suite';
+import { useDevice } from 'src/hooks/suite';
 import { useFirmwareDesktopUpdate } from 'src/hooks/suite/useFirmwareDesktopUpdate';
-import { selectIsDebugModeActive } from 'src/selectors/suite/suiteSelectors';
-
-import { PrerequisitesGuide } from '../suite';
-import { FirmwareLowBatteryModal } from './FirmwareLowBatteryModal';
-import { FirmwareSwitchWarning } from './FirmwareSwitchWarning';
-
-const TextButton = styled.button`
-    background: none;
-    border: none;
-    cursor: pointer;
-    font-size: inherit;
-    padding: 0;
-    text-decoration: underline;
-`;
-
-const WarningListWrapper = styled.div`
-    display: flex;
-    align-items: flex-start;
-    flex-direction: column;
-    gap: ${spacingsPx.md};
-    border-bottom: 1px solid ${({ theme }) => theme.legacy.STROKE_GREY};
-    margin: ${spacingsPx.xs} ${spacingsPx.md};
-    padding-bottom: ${spacingsPx.md};
-`;
-
-const Important = styled.div`
-    align-self: flex-start;
-    color: ${({ theme }) => theme.legacy.TYPE_LIGHT_GREY};
-    font-weight: ${variables.FONT_WEIGHT.MEDIUM};
-    text-transform: uppercase;
-`;
-
-const EmphasizedText = styled.b`
-    color: ${({ theme }) => theme.legacy.TYPE_DARK_GREY};
-    font-weight: ${variables.FONT_WEIGHT.DEMI_BOLD};
-`;
-
-const InstallButton = ({ isDisabled, onClick, variant, children }: ButtonProps) => (
-    <Tooltip
-        cursor="default"
-        isActive={isDisabled}
-        maxWidth={200}
-        placement="bottom"
-        content={<Translation id="TR_INSTALL_FW_DISABLED_MULTIPLE_DEVICES" />}
-    >
-        <OnboardingCard.Button
-            variant={variant}
-            onClick={onClick}
-            isDisabled={isDisabled}
-            data-testid="@firmware/install-button"
-        >
-            {children}
-        </OnboardingCard.Button>
-    </Tooltip>
-);
 
 type GetDescriptionProps = {
     required: boolean;
+    reinstall: boolean;
     targetType: FirmwareType;
     shouldSwitchFirmwareType?: boolean;
     isBitcoinOnlyAvailable?: boolean;
@@ -81,6 +18,7 @@ type GetDescriptionProps = {
 
 const getDescription = ({
     required,
+    reinstall,
     targetType,
     shouldSwitchFirmwareType,
     isBitcoinOnlyAvailable,
@@ -99,211 +37,37 @@ const getDescription = ({
         return 'TR_FIRMWARE_UPDATE_REQUIRED_EXPLAINED';
     }
 
-    return 'TR_ONBOARDING_NEW_FW_DESCRIPTION';
-};
-
-const getNoFirmwareInstalledSubheading = (device: AcquiredDevice) => {
-    const bitcoinOnlyDevice = isBitcoinOnlyDevice(device);
-
-    if (bitcoinOnlyDevice) {
-        return device.firmware === 'none'
-            ? 'TR_FIRMWARE_SUBHEADING_NONE_BITCOIN_ONLY_DEVICE'
-            : 'TR_FIRMWARE_SUBHEADING_UNKNOWN_BITCOIN_ONLY_DEVICE';
-    }
-
-    return device.firmware === 'none'
-        ? 'TR_FIRMWARE_SUBHEADING_NONE'
-        : 'TR_FIRMWARE_SUBHEADING_UNKNOWN';
+    return reinstall ? 'TR_FIRMWARE_REINSTALL_FW_DESCRIPTION' : 'TR_FIRMWARE_NEW_FW_DESCRIPTION';
 };
 
 type FirmwareInitialProps = {
     shouldSwitchFirmwareType?: boolean;
-    // This component is shared between Onboarding flow and standalone fw update modal with few minor UI changes
-    // If it is set to true, then you know it is being rendered in standalone fw update modal
-    onClose?: () => void;
 };
 
-// TODO: consolidate with FirmwareInitialStandalone
-export const FirmwareInitial = ({
-    shouldSwitchFirmwareType = false,
-    onClose,
-}: FirmwareInitialProps) => {
+export const FirmwareInitial = ({ shouldSwitchFirmwareType = false }: FirmwareInitialProps) => {
     const { device } = useDevice();
-    const {
-        deviceWillBeWiped,
-        firmwareUpdate,
-        setStatus,
-        targetFirmwareType,
-        toggleLowBatteryModal,
-        showLowBatteryModal,
-    } = useFirmwareDesktopUpdate({
+    const { deviceWillBeWiped, targetFirmwareType } = useFirmwareDesktopUpdate({
         shouldSwitchFirmwareType,
     });
-    const { isActive: isOnboarding, updateAnalytics } = useOnboarding();
-    const { translationString } = useTranslation();
-    const devices = useSelector(selectDevices);
-    const isDebug = useSelector(selectIsDebugModeActive);
-
-    const [bitcoinOnlyOffer, setBitcoinOnlyOffer] = useState(false);
-    const [showSkipConfirmation, setShowSkipConfirmation] = useState(false);
 
     // Just to satisfy TS, disconnected device should be handled upstream.
     if (!device?.connected || !device?.features) {
         return null;
     }
 
-    // This is essential for detecting potentially malicious devices,
-    // as the firmware revision check is a critical part of the verification process.
-    // In debug mode you may bypass it.
-    // See: https://github.com/trezor/trezor-suite/issues/19157
-    const isFirmwareInstallationMandatory =
-        device.features.internal_model === DeviceModelInternal.T1B1 && !isDebug;
-
-    // todo: move to utils device.ts
-    const devicesConnected = devices.filter(device => device?.connected);
-    const multipleDevicesConnected = [...new Set(devicesConnected.map(d => d.path))].length > 1;
-
-    // The first condition is a defensive measure against https://github.com/trezor/trezor-suite/issues/17246, I could not reproduce the error.
-    const shouldCheckSeed = !isOnboarding && device?.mode !== 'initialize';
-
-    let content;
-
-    const targetType = bitcoinOnlyOffer ? FirmwareType.BitcoinOnly : targetFirmwareType;
     // Bitcoin-only firmware is only available on T2T1 from v2.0.8 - older devices must first upgrade to 2.1.1 which does not have a Bitcoin-only variant
     const isBitcoinOnlyAvailable = !!device.firmwareReleaseConfigInfo?.isBitcoinOnlyAvailable;
+    const currentFwVersion = getFirmwareVersion(device);
+    const availableFwVersion = getFwUpdateVersion(device);
+    const hasLatestAvailableFw = !!(
+        availableFwVersion &&
+        currentFwVersion &&
+        availableFwVersion === currentFwVersion
+    );
 
-    const installFirmware = (firmwareType: FirmwareType) => {
-        firmwareUpdate({ firmwareType });
-        updateAnalytics({ firmware: 'install' });
-    };
-
-    if (showLowBatteryModal) {
-        return <FirmwareLowBatteryModal onClose={toggleLowBatteryModal} />;
-    }
-
-    if (bitcoinOnlyOffer) {
-        // Installing Bitcoin-only firmware in onboarding
-        content = {
-            heading: (
-                <Translation
-                    id="TR_INSTALL_BITCOIN_ONLY_FW"
-                    values={{
-                        bitcoinOnly: <Translation id="TR_FIRMWARE_TYPE_BITCOIN_ONLY" />,
-                    }}
-                />
-            ),
-            description: (
-                <Column alignItems="center" gap={12}>
-                    <Translation id="TR_FIRMWARE_SUBHEADING_BITCOIN" />
-                    <Note>
-                        <Translation id="TR_CHANGE_FIRMWARE_TYPE_ANYTIME" />
-                    </Note>
-                </Column>
-            ),
-            body: (
-                <Column gap={32} alignItems="center">
-                    <FirmwareOffer targetFirmwareType={FirmwareType.BitcoinOnly} />
-                    <FirmwareWarningsList />
-                </Column>
-            ),
-            innerActions: (
-                <Row gap={16}>
-                    <InstallButton
-                        variant="tertiary"
-                        onClick={() => installFirmware(FirmwareType.Universal)}
-                        isDisabled={multipleDevicesConnected}
-                    >
-                        <Translation
-                            id="TR_INSTALL_REGULAR"
-                            values={{
-                                regular: <Translation id="TR_FIRMWARE_TYPE_REGULAR" />,
-                            }}
-                        />
-                    </InstallButton>
-                    <InstallButton
-                        onClick={() => installFirmware(targetType)}
-                        isDisabled={multipleDevicesConnected}
-                    >
-                        <Translation
-                            id="TR_INSTALL_BITCOIN_ONLY"
-                            values={{
-                                bitcoinOnly: <Translation id="TR_FIRMWARE_TYPE_BITCOIN_ONLY" />,
-                            }}
-                        />
-                    </InstallButton>
-                </Row>
-            ),
-        };
-    } else if (['none', 'unknown'].includes(device.firmware)) {
-        const subheadingId = getNoFirmwareInstalledSubheading(device);
-
-        // No firmware installed
-        // Device without firmware is already in bootloader mode even if it doesn't report it
-        content = {
-            heading: <Translation id="TR_INSTALL_FIRMWARE" />,
-            description: (
-                <Translation
-                    id={subheadingId}
-                    values={{
-                        i: chunks => <i>{chunks}</i>,
-                        button: chunks => (
-                            <TextButton onClick={() => setBitcoinOnlyOffer(true)}>
-                                {chunks}
-                            </TextButton>
-                        ),
-                        bitcoinOnly: translationString('TR_FIRMWARE_TYPE_BITCOIN_ONLY'),
-                    }}
-                />
-            ),
-            body: (
-                <Column gap={32} alignItems="center">
-                    <FirmwareOffer targetFirmwareType={targetType} />
-                    <FirmwareWarningsList />
-                </Column>
-            ),
-            innerActions: (
-                <InstallButton
-                    onClick={() => installFirmware(targetType)}
-                    isDisabled={multipleDevicesConnected}
-                >
-                    <Translation id="TR_INSTALL" />
-                </InstallButton>
-            ),
-        };
-    } else if (device.mode === 'bootloader') {
-        // We can check if device.mode is bootloader only after checking that firmware !== none (condition above)
-        // because device without firmware always reports that it is in bootloader mode.
-        //
-        // We want to prevent FW installation directly from bootloader only during onboarding,
-        // because we want to read current FW version from the device first and cache it.
-        // But for standalone FW update we need to allow bootloader mode directly, because
-        // the device could be stucked in bootloader (e.g. wrong intermediary FW installation).
-        return <PrerequisitesGuide />;
-    } else if (device.firmware === 'required' || device.firmware === 'outdated') {
-        const warningTranslationValues: ExtendedMessageDescriptor['values'] = {
-            b: chunks => <EmphasizedText>{chunks}</EmphasizedText>,
-        };
-
-        content = {
-            heading: shouldSwitchFirmwareType ? (
-                <Translation
-                    id="TR_SWITCH_FIRMWARE_TO"
-                    values={{
-                        firmwareType: (
-                            <Translation
-                                id={
-                                    targetType === FirmwareType.BitcoinOnly
-                                        ? 'TR_FIRMWARE_TYPE_BITCOIN_ONLY'
-                                        : 'TR_FIRMWARE_TYPE_REGULAR'
-                                }
-                            />
-                        ),
-                    }}
-                />
-            ) : (
-                <Translation id="TR_INSTALL_FIRMWARE" />
-            ),
-            description: (
+    return (
+        <Column gap={16}>
+            <Banner variant="info" icon="info">
                 <Translation
                     id={getDescription({
                         /**
@@ -315,7 +79,8 @@ export const FirmwareInitial = ({
                          *   so it should not be used here.
                          */
                         required: device.firmware === 'required',
-                        targetType,
+                        reinstall: device.firmware === 'valid' || hasLatestAvailableFw,
+                        targetType: targetFirmwareType,
                         shouldSwitchFirmwareType,
                         isBitcoinOnlyAvailable,
                     })}
@@ -324,83 +89,12 @@ export const FirmwareInitial = ({
                         regular: <Translation id="TR_FIRMWARE_TYPE_REGULAR" />,
                     }}
                 />
-            ),
-            body: (
-                <Column gap={32} alignItems="center">
-                    {deviceWillBeWiped && (
-                        <WarningListWrapper>
-                            <Important>
-                                <Translation id="TR_IMPORTANT" />
-                            </Important>
-                            <FirmwareSwitchWarning>
-                                <Translation
-                                    id="TR_FIRMWARE_SWITCH_WARNING_1"
-                                    values={warningTranslationValues}
-                                />
-                            </FirmwareSwitchWarning>
-                            <FirmwareSwitchWarning>
-                                <Translation
-                                    id="TR_FIRMWARE_SWITCH_WARNING_2"
-                                    values={warningTranslationValues}
-                                />
-                            </FirmwareSwitchWarning>
-                        </WarningListWrapper>
-                    )}
-                    <FirmwareOffer targetFirmwareType={targetType} />
-                    <FirmwareWarningsList />
-                </Column>
-            ),
-            innerActions: (
-                <Row gap={16}>
-                    <InstallButton
-                        onClick={() =>
-                            shouldCheckSeed ? setStatus('check-seed') : installFirmware(targetType)
-                        }
-                        isDisabled={multipleDevicesConnected}
-                    >
-                        <Translation id={deviceWillBeWiped ? 'TR_CONTINUE' : 'TR_INSTALL'} />
-                    </InstallButton>
-                    {deviceWillBeWiped && onClose && (
-                        <OnboardingCard.Button onClick={onClose} variant="tertiary">
-                            <Translation id="TR_CLOSE" />
-                        </OnboardingCard.Button>
-                    )}
-                </Row>
-            ),
-            outerActions:
-                device.firmware === 'outdated' && !isFirmwareInstallationMandatory ? (
-                    <OnboardingCard.SecondaryButton
-                        onClick={() => {
-                            setShowSkipConfirmation(true);
-                            updateAnalytics({ firmware: 'skip' });
-                        }}
-                        data-testid="@firmware/skip-button"
-                    >
-                        <Translation id="TR_SKIP_UPDATE" />
-                    </OnboardingCard.SecondaryButton>
-                ) : undefined,
-        };
-    }
-
-    if (content) {
-        return (
-            <>
-                {showSkipConfirmation && (
-                    <SkipStepConfirmation onCancel={() => setShowSkipConfirmation(false)} />
-                )}
-                <OnboardingCard
-                    iconName="circuitry"
-                    heading={content.heading}
-                    description={content.description}
-                    innerActions={content.innerActions}
-                    outerActions={content.outerActions}
-                    isActionAbortable={false}
-                >
-                    {content.body}
-                </OnboardingCard>
-            </>
-        );
-    }
-
-    return null;
+            </Banner>
+            {deviceWillBeWiped && <FirmwareWipeWarning />}
+            <Card>
+                <FirmwareOffer targetFirmwareType={targetFirmwareType} />
+            </Card>
+            <FirmwareWarningsList />
+        </Column>
+    );
 };
