+++ packages/suite/src/components/suite/SecurityCheck/SecurityCheckFail.tsx
@@ -5,8 +5,8 @@ import { Column, Divider, H2, Paragraph, Row } from '@trezor/components';
 import { spacings } from '@trezor/theme';
 
 import { Translation } from 'src/components/suite/Translation';
-import { SecurityChecklist } from 'src/views/onboarding/steps/SecurityCheck/SecurityChecklist';
-import { SecurityChecklistItem } from 'src/views/onboarding/steps/SecurityCheck/types';
+import { SecurityChecklist } from 'src/views/onboarding/steps/DeviceAuthenticityStep/SecurityChecklist';
+import { SecurityChecklistItem } from 'src/views/onboarding/steps/DeviceAuthenticityStep/types';
 
 import { SecurityCheckLayout } from './SecurityCheckLayout';
 import { hardFailureChecklistItems } from './checklistItems';
