+++ packages/suite/src/views/onboarding/UnexpectedState/DeviceDisconnectedStep.tsx
