+++ packages/suite/src/views/onboarding/steps/CreateOrRecoverStep.tsx
