+++ packages/suite/src/views/onboarding/steps/DeviceAuthenticityStep/SecurityChecklist.tsx
