+++ packages/suite/src/views/onboarding/steps/DeviceAuthenticityStep/__tests__/types.test.tsx
