+++ packages/suite/src/views/onboarding/steps/DeviceAuthenticityStep/types.ts
