+++ packages/suite/src/views/onboarding/steps/SecurityStep.tsx
